Cleaned up the master branch to try and get a fresh start in this repository.
Converted floodlight.py to work with the EAGER version of Floodlight and got rid of all of the git root directory stuff. Now Floodlight is just being installed in the parent directory.

Also removed context.py. Apparently sibling modules in python don't work well together, so for transparency, all the path changes done to import siblings are just done in the headers of the test modules.